requirements.txt
asgiref==3.3.1
backcall==0.2.0
decorator==4.4.2
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
Markdown==3.3.3
numpy==1.20.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
pytesseract==0.2.5
python-dateutil==2.7.3
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
